+++ src/Exceptions/WrongFollowingType.php
@@ -4,5 +4,4 @@ namespace seregazhuk\PinterestBot\Exceptions;
 
 class WrongFollowingType extends PinterestBotException
 {
-
 }