+++ src/Api/Request.php
@@ -2,11 +2,11 @@
 
 namespace seregazhuk\PinterestBot\Api;
 
-use seregazhuk\PinterestBot\Helpers\Cookies;
-use seregazhuk\PinterestBot\Helpers\UrlBuilder;
-use seregazhuk\PinterestBot\Helpers\FileHelper;
 use seregazhuk\PinterestBot\Api\Contracts\HttpClient;
 use seregazhuk\PinterestBot\Exceptions\InvalidRequest;
+use seregazhuk\PinterestBot\Helpers\Cookies;
+use seregazhuk\PinterestBot\Helpers\FileHelper;
+use seregazhuk\PinterestBot\Helpers\UrlBuilder;
 
 /**
  * Class Request.
@@ -194,6 +194,7 @@ class Request
         $request = self::createRequestData(
             ['options' => $data], $bookmarks
         );
+        print_r($request);
 
         return UrlBuilder::buildRequestString($request);
     }
@@ -218,7 +219,7 @@ class Request
 
         return [
             'source_url' => '',
-            'data'       => json_encode($data),
+            'data'       => json_encode($data, JSON_FORCE_OBJECT),
         ];
     }